refactor(settings): changes asked in #22 pr
