+++ assets/scss/_settings.scss
@@ -36,7 +36,7 @@
     border: 0.15741vh solid rgba(255, 255, 255, 0.2);
     height: calc(100% - 11.703703703703704vh);
     background: rgba(0, 0, 0, 0.8);
-    backdrop-filter: blur(10px);
+    backdrop-filter: blur(0.9259259259259vh);
   }
 
   h2 {
